+++ lib/src/in_app_webview/main.dart
@@ -4,4 +4,4 @@ export 'in_app_webview_controller.dart';
 export 'in_app_webview_settings.dart';
 export 'headless_in_app_webview.dart';
 export 'android/main.dart';
-export 'ios/main.dart';
+export 'apple/main.dart';