+++ lib/src/main.dart
@@ -2,7 +2,7 @@ export 'in_app_webview/main.dart';
 export 'in_app_browser/main.dart';
 export 'chrome_safari_browser/main.dart';
 export 'android/main.dart';
-export 'ios/main.dart';
+export 'apple/main.dart';
 export 'x509_certificate/main.dart';
 export 'web_storage/main.dart';
 export 'types.dart';